requirements.txt
Pillow>=11.3.0
streamlit>=1.45.1
torch>=2.7.1+cu118
torchvision>=0.22.1+cu118
